requirements.txt
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
transformers==4.54.1
pillow==11.3.0
